+++ application-portal/api/.sequelizerc
@@ -1,0 +1,6 @@
+{
+  "config": "src/database/config.js",
+  "migrations-path": "src/database/migrations",
+  "models-path": "src/modules",
+  "seeders-path": "src/database/seeders"
+}
