+++ application-portal/api/package.json
@@ -11,6 +11,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@apollo/client": "^3.13.5",
     "@apollo/server": "^4.11.3",
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
