+++ application-portal/api/src/database/config.js
@@ -3,7 +3,7 @@ import 'dotenv/config'
 export const config = {
   development: {
     dialect: 'sqlite',
-    storage: './src/database.sqlite',
+    storage: './src/database/database.sqlite',
   },
   test: {
     dialect: 'sqlite',
@@ -11,6 +11,6 @@ export const config = {
   },
   production: {
     dialect: 'sqlite',
-    storage: './src/database.sqlite',
+    storage: './src/database/database.sqlite',
   },
 }
