--- application-portal/api/src/models/Company.js
@@ -1,17 +1,0 @@
-import { DataTypes } from 'sequelize'
-
-export default (sequelize) => {
-  const Company = sequelize.define('Company', {
-    id: {
-      type: DataTypes.INTEGER,
-      autoIncrement: true,
-      allowNull: false,
-      primaryKey: true,
-    },
-    name: DataTypes.STRING,
-    description: DataTypes.STRING,
-    contactEmail: DataTypes.STRING,
-    contactPhone: DataTypes.STRING,
-  })
-  return Company
-}
