+++ application-portal/api/src/modules/application/application.model.js
@@ -1,0 +1,40 @@
+import { DataTypes } from 'sequelize'
+
+export default (sequelize) => {
+  const Application = sequelize.define('Application', {
+    id: {
+      type: DataTypes.INTEGER,
+      autoIncrement: true,
+      allowNull: false,
+      primaryKey: true,
+    },
+    title: DataTypes.STRING,
+    type: DataTypes.STRING,
+    status: {
+      type: DataTypes.STRING,
+      defaultValue: 'Pending',
+    },
+    description: DataTypes.STRING,
+    location: DataTypes.STRING,
+    latlong: DataTypes.JSON,
+    minSalary: DataTypes.INTEGER,
+    maxSalary: DataTypes.INTEGER,
+    link: DataTypes.STRING,
+    skills: DataTypes.JSON,
+    reminderEmailSent: {
+      type: DataTypes.BOOLEAN,
+      allowNull: false,
+      defaultValue: false,
+    },
+    CompanyId: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      references: {
+        model: 'Companies',
+        key: 'id',
+      },
+    },
+  })
+
+  return Application
+}
