+++ application-portal/api/src/modules/company/company.service.js
@@ -1,4 +1,4 @@
-import db from '../../models/index.js'
+import db from '../../database/initDb.js'
 
 const { Application, Company } = db
